Server/requirements.txt
fastapi>=0.95.2
uvicorn[standard]>=0.23.1
python-dotenv>=1.0.0
redis>=4.5.4
bcrypt>=4.0.1
pycryptodome>=3.18.0
